+++ webdriver/tests/support/fixtures.py
@@ -5,7 +5,6 @@ from urllib.parse import urlunsplit
 
 from tests.support.helpers import is_wayland
 from tests.support.web_extension import EXTENSION_DATA
-from tests.support.inline import build_inline
 from tests.support.http_request import HTTPRequest
 from tests.support.keys import Keys
 
@@ -105,34 +104,6 @@ def modifier_key(current_session):
         return Keys.META
     else:
         return Keys.CONTROL
-
-
-@pytest.fixture
-def inline(url):
-    """Take a source extract and produces well-formed documents.
-
-    Based on the desired document type, the extract is embedded with
-    predefined boilerplate in order to produce well-formed documents.
-    The media type and character set may also be individually configured.
-
-    This helper function originally used data URLs, but since these
-    are not universally supported (or indeed standardised!) across
-    browsers, it now delegates the serving of the document to wptserve.
-    This file also acts as a wptserve handler (see the main function
-    below) which configures the HTTP response using query parameters.
-
-    This function returns a URL to the wptserve handler, which in turn
-    will serve an HTTP response with the requested source extract
-    inlined in a well-formed document, and the Content-Type header
-    optionally configured using the desired media type and character set.
-
-    Any additional keyword arguments are passed on to the build_url
-    function, which comes from the url fixture.
-    """
-    def inline(src, **kwargs):
-        return build_inline(url, src, **kwargs)
-
-    return inline
 
 
 @pytest.fixture
